refactor(PieceFactory): 팩토리 메서드 분리
