+++ src/main/java/janggi/presentation/dto/PositionInfo.java
@@ -2,6 +2,7 @@ package janggi.presentation.dto;
 
 import janggi.domain.Point;
 import janggi.domain.piece.Piece;
+import janggi.domain.piece.PieceFactory;
 import janggi.domain.piece.PieceType;
 import janggi.domain.status.Team;
 
@@ -11,7 +12,7 @@ public record PositionInfo(
 ) {
     public static PositionInfo from(Team team, String pieceName, int x, int y) {
         return new PositionInfo(
-                PieceType.createPiece(team, PieceType.valueOf(pieceName)),
+                PieceFactory.createPiece(team, PieceType.valueOf(pieceName)),
                 Point.of(x, y)
         );
     }
